timelog_backend/pyproject.toml
[tool.poetry]
name = "timelog-backend"
version = "0.1.0"
description = ""
authors = ["dapt <diegodapt18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.5"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
